+++ core/include/helper.h
@@ -31,8 +31,8 @@ public:
     static std::vector<std::string> splitString(const std::string& s, char delimiter);
     static const std::string generateIdentation(int identPosition, int identSize = 2);
 
-    static const std::string ticks2noteType(const size_t ticks, const size_t divisionsPerQuarterNote);
-    static size_t noteType2ticks(const std::string& noteType, const size_t divisionsPerQuarterNote);
+    static const std::string ticks2noteType(const int ticks, const int divisionsPerQuarterNote = 256);
+    static int noteType2ticks(const std::string& noteType, const int divisionsPerQuarterNote = 256);
 
     static float ticks2QuarterNoteValue(const size_t ticks, const size_t divisionsPerQuarterNote);