+++ core/src/python_wrapper/py_helper.cpp
@@ -28,6 +28,17 @@ void HelperClass(py::module &m) {
         py::arg("accType") = std::string(), 
         py::call_guard<py::scoped_ostream_redirect, py::scoped_estream_redirect>());
     //--------------------- //
+
+    cls.def_static("noteType2ticks", &Helper::noteType2ticks, 
+        py::arg("noteType"),
+        py::arg("divisionsPerQuarterNote") = 256,
+        py::call_guard<py::scoped_ostream_redirect, py::scoped_estream_redirect>());
+
+    cls.def_static("ticks2noteType", &Helper::ticks2noteType, 
+        py::arg("ticks"),
+        py::arg("divisionsPerQuarterNote") = 256,
+        py::call_guard<py::scoped_ostream_redirect, py::scoped_estream_redirect>());
+
     cls.def_static("pitch2number", &Helper::pitch2number, 
         py::call_guard<py::scoped_ostream_redirect, py::scoped_estream_redirect>());
     //--------------------- //